# openapi.yaml
openapi: 3.1.0
info:
  title: MinusPod API
  description: |
    REST API for managing podcast feeds with automatic ad detection and removal.


    This API allows you to:
    - Manage podcast feed subscriptions
    - View and manage processed episodes
    - Configure ad detection settings including AI model selection
    - Monitor system status and trigger cleanup operations
    - Manage cross-episode ad patterns with network and podcast scope
    - Submit corrections to improve ad detection accuracy
  version: 1.0.64
  contact:
    name: MinusPod
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: '{baseUrl}/api/v1'
    description: Custom server
    variables:
      baseUrl:
        default: http://localhost:8000
        description: Base URL of the podcast server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login with password
      description: Authenticate with the configured password. Rate limited to 5 attempts per minute.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  description: The application password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /feeds:
    get:
      tags:
        - Feeds
      summary: List all feeds
      description: Returns a list of all configured podcast feeds with metadata
      operationId: listFeeds
      responses:
        '200':
          description: List of feeds
          content:
            application/json:
              schema:
                type: object
                properties:
                  feeds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Feed'
    post:
      tags:
        - Feeds
      summary: Add a new feed
      description: Add a new podcast feed by providing the source RSS URL
      operationId: addFeed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceUrl
              properties:
                sourceUrl:
                  type: string
                  format: uri
                  description: URL of the source RSS feed
                  example: https://feeds.example.com/podcast.xml
                slug:
                  type: string
                  description: Optional custom slug (auto-generated if not provided)
                  example: my-podcast
                maxEpisodes:
                  type: integer
                  minimum: 10
                  maximum: 500
                  description: Maximum number of episodes served in the RSS feed (default 300)
      responses:
        '201':
          description: Feed created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  slug:
                    type: string
                  sourceUrl:
                    type: string
                  feedUrl:
                    type: string
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Feed with slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  schemas:
    Feed:
      type: object
      properties:
        slug:
          type: string
          description: URL-safe identifier for the feed
        title:
          type: string
          description: Podcast title from RSS feed
        sourceUrl:
          type: string
          format: uri
          description: Original RSS feed URL
        feedUrl:
          type: string
          format: uri
          description: Modified RSS feed URL served by this server
        artworkUrl:
          type:
            - string
            - 'null'
          description: URL to cached artwork image
        episodeCount:
          type: integer
          description: Total number of episodes
        processedCount:
          type: integer
          description: Number of processed episodes
        lastRefreshed:
          type:
            - string
            - 'null'
          format: date-time
          description: Last RSS refresh timestamp
        createdAt:
          type: string
          format: date-time
        autoProcessOverride:
          type:
            - boolean
            - 'null'
          description: Override auto-process setting for this feed (null uses global setting)
        maxEpisodes:
          type:
            - integer
            - 'null'
          minimum: 10
          maximum: 500
          description: Max episodes served to podcast clients via modified RSS. Null uses default (300).
...
    Error:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...